+++ test/snippet/core/algorithm/configuration_combine.cpp
@@ -1,4 +1,3 @@
-#include <seqan3/alignment/band/static_band.hpp>
 #include <seqan3/alignment/configuration/align_config_band.hpp>
 #include <seqan3/alignment/configuration/align_config_mode.hpp>
 #include <seqan3/core/algorithm/configuration.hpp>
@@ -8,6 +7,6 @@ int main()
 {
     // Here we use the global and banded alignment configurations to show how they can be combined.
     seqan3::configuration my_cfg = seqan3::align_cfg::mode{seqan3::global_alignment} |
-                                   seqan3::align_cfg::band{seqan3::static_band{seqan3::lower_bound{-4},
-                                                                               seqan3::upper_bound{4}}};
+                                   seqan3::align_cfg::band_fixed_size{seqan3::align_cfg::lower_diagonal{-4},
+                                                                      seqan3::align_cfg::upper_diagonal{4}};
 }